+++ src/test/java/uk/gov/hmcts/reform/finrem/caseorchestration/utils/retry/RetryExecutorTest.java
@@ -1,39 +1,21 @@
 package uk.gov.hmcts.reform.finrem.caseorchestration.utils.retry;
 
-import feign.FeignException;
 import org.junit.jupiter.api.AfterEach;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.Nested;
 import org.junit.jupiter.api.Test;
 import org.junit.jupiter.api.extension.ExtendWith;
-import org.junit.jupiter.params.ParameterizedTest;
-import org.junit.jupiter.params.provider.ValueSource;
 import org.mockito.ArgumentCaptor;
 import org.mockito.Mock;
 import org.mockito.MockedStatic;
 import org.mockito.Spy;
 import org.mockito.junit.jupiter.MockitoExtension;
 import org.springframework.aop.framework.AopContext;
-import org.springframework.beans.factory.annotation.Autowired;
-import org.springframework.context.annotation.Bean;
-import org.springframework.context.annotation.Configuration;
-import org.springframework.context.annotation.Import;
-import org.springframework.retry.RetryListener;
-import org.springframework.retry.annotation.EnableRetry;
-import org.springframework.retry.backoff.FixedBackOffPolicy;
-import org.springframework.retry.interceptor.RetryInterceptorBuilder;
-import org.springframework.retry.interceptor.RetryOperationsInterceptor;
-import org.springframework.retry.policy.SimpleRetryPolicy;
-import org.springframework.retry.support.RetryTemplate;
-import org.springframework.test.context.junit.jupiter.SpringJUnitConfig;
 import org.springframework.util.function.ThrowingSupplier;
 import uk.gov.hmcts.reform.finrem.caseorchestration.util.TestLogger;
 import uk.gov.hmcts.reform.finrem.caseorchestration.util.TestLogs;
 
-import java.util.HashMap;
-import java.util.Map;
 import java.util.Optional;
-import java.util.concurrent.atomic.AtomicInteger;
 
 import static org.assertj.core.api.Assertions.assertThat;
 import static org.assertj.core.api.Assertions.assertThatThrownBy;
@@ -46,47 +28,27 @@ import static org.mockito.Mockito.verify;
 import static org.mockito.Mockito.verifyNoInteractions;
 import static org.mockito.Mockito.verifyNoMoreInteractions;
 import static uk.gov.hmcts.reform.finrem.caseorchestration.TestConstants.CASE_ID;
-import static uk.gov.hmcts.reform.finrem.caseorchestration.TestSetUpUtils.feignException;
 import static uk.gov.hmcts.reform.finrem.caseorchestration.TestSetUpUtils.getThrowingSupplierCaptor;
 
-@SpringJUnitConfig(RetryExecutorTest.Config.class)
 @ExtendWith(MockitoExtension.class)
 class RetryExecutorTest {
     @TestLogs
     private final TestLogger logs = new TestLogger(RetryExecutor.class);
 
-    private static final int DEFAULT_MAX_ATTEMPTS = 3;
+    @Spy
+    RetryExecutor spyExecutor;
+    private MockedStatic<AopContext> aopContextMock;
 
-    @Autowired
-    private RetryExecutor retryExecutor;
+    @BeforeEach
+    void setUp() {
+        aopContextMock = mockStatic(AopContext.class);
+        aopContextMock.when(AopContext::currentProxy)
+            .thenReturn(spyExecutor);
+    }
 
-    @Configuration
-    @EnableRetry
-    @Import(RetryExecutor.class)
-    static class Config {
-
-        @Bean(name = "retryLoggerInterceptor")
-        public RetryOperationsInterceptor retryLoggerInterceptor() {
-            Map<Class<? extends Throwable>, Boolean> retryableExceptions = new HashMap<>();
-            retryableExceptions.put(FeignException.InternalServerError.class, true);
-            retryableExceptions.put(FeignException.ServiceUnavailable.class, true);
-            retryableExceptions.put(FeignException.GatewayTimeout.class, true);
-
-            SimpleRetryPolicy retryPolicy = new SimpleRetryPolicy(3, retryableExceptions);
-
-            RetryTemplate retryTemplate = new RetryTemplate();
-            retryTemplate.setRetryPolicy(retryPolicy);
-
-            FixedBackOffPolicy backOffPolicy = new FixedBackOffPolicy();
-            backOffPolicy.setBackOffPeriod(2000);
-            retryTemplate.setBackOffPolicy(backOffPolicy);
-
-            retryTemplate.setListeners(new RetryListener[]{ new RetryLogger() });
-
-            return RetryInterceptorBuilder.stateless()
-                .retryOperations(retryTemplate)
-                .build();
-        }
+    @AfterEach
+    void tearDown() {
+        aopContextMock.close();
     }
 
     @Nested
@@ -96,21 +58,6 @@ class RetryExecutorTest {
         private RetryErrorHandler handler1;
         @Mock
         private RetryErrorHandler handler2;
-        @Spy
-        RetryExecutor spyExecutor;
-        private MockedStatic<AopContext> aopContextMock;
-
-        @BeforeEach
-        void setUp() {
-            aopContextMock = mockStatic(AopContext.class);
-            aopContextMock.when(AopContext::currentProxy)
-                .thenReturn(spyExecutor);
-        }
-
-        @AfterEach
-        void tearDown() {
-            aopContextMock.close(); // IMPORTANT
-        }
 
         @Test
         void shouldExecuteSuccessfully_whenNoExceptionThrown() {
@@ -166,10 +113,6 @@ class RetryExecutorTest {
             ThrowingRunnable action = mock(ThrowingRunnable.class);
             Exception exception = new RuntimeException("boom");
 
-            doThrow(exception)
-                .when(spyExecutor)
-                .runWithRetry(action, "testAction", CASE_ID);
-
             assertThatThrownBy(() ->
                 spyExecutor.runWithRetryWithHandler(action, "testAction", CASE_ID)
             )
@@ -203,21 +146,6 @@ class RetryExecutorTest {
         private RetryErrorHandler handler1;
         @Mock
         private RetryErrorHandler handler2;
-        @Spy
-        RetryExecutor spyExecutor;
-        private MockedStatic<AopContext> aopContextMock;
-
-        @BeforeEach
-        void setUp() {
-            aopContextMock = mockStatic(AopContext.class);
-            aopContextMock.when(AopContext::currentProxy)
-                .thenReturn(spyExecutor);
-        }
-
-        @AfterEach
-        void tearDown() {
-            aopContextMock.close(); // IMPORTANT
-        }
 
         @Test
         void shouldExecuteSuccessfully_whenNoExceptionThrown() {
@@ -275,13 +203,8 @@ class RetryExecutorTest {
         }
 
         @Test
-        void shouldThrowIllegalStateException_whenNoHandlerProvided() throws Exception {
+        void shouldThrowIllegalStateException_whenNoHandlerProvided() {
             ThrowingSupplier<?> action = mock(ThrowingSupplier.class);
-            Exception exception = new RuntimeException("boom");
-
-            doThrow(exception)
-                .when(spyExecutor)
-                .supplyWithRetry(action, "testAction", CASE_ID);
 
             assertThatThrownBy(() ->
                 spyExecutor.supplyWithRetryWithHandler(action, "testAction", CASE_ID)
@@ -308,102 +231,6 @@ class RetryExecutorTest {
                         + "This indicates a bug or retry exhausted in RetryExecutor.").formatted(CASE_ID))
             );
 
-        }
-    }
-
-    @Nested
-    class RunWithRetryTest {
-
-        @ParameterizedTest(name = "should retry on status {0} and succeed")
-        @ValueSource(ints = {500, 503, 504})
-        void shouldRetryOnRetryableStatus(int status) throws Exception {
-            AtomicInteger counter = new AtomicInteger();
-
-            retryExecutor.runWithRetry(() -> {
-                if (counter.getAndIncrement() < DEFAULT_MAX_ATTEMPTS - 1) {
-                    throw feignException(status, "Retryable Error");
-                }
-            }, "send", CASE_ID);
-
-            assertThat(counter.get()).isEqualTo(DEFAULT_MAX_ATTEMPTS);
-        }
-
-        @ParameterizedTest(name = "should retry once on first status {0} then succeed")
-        @ValueSource(ints = {504}) // can expand if needed
-        void shouldRetryOnceThenSucceed(int status) throws Exception {
-            AtomicInteger counter = new AtomicInteger();
-
-            retryExecutor.runWithRetry(() -> {
-                if (counter.getAndIncrement() == 0) {
-                    throw feignException(status, "Retryable Error");
-                }
-            }, "send", CASE_ID);
-
-            assertThat(counter.get()).isEqualTo(2);
-        }
-
-        @Test
-        void shouldNotRetryOnNonRetryableException() {
-            AtomicInteger counter = new AtomicInteger();
-
-            assertThatThrownBy(() ->
-                retryExecutor.supplyWithRetry(() -> {
-                    counter.incrementAndGet();
-                    throw new RuntimeException("Not retryable");
-                }, "send", CASE_ID)
-            ).isInstanceOf(RuntimeException.class);
-
-            assertThat(counter.get()).isEqualTo(1);
-        }
-    }
-
-    @Nested
-    class SupplyWithRetryTest {
-
-        @ParameterizedTest(name = "should retry on status {0} and succeed")
-        @ValueSource(ints = {500, 503, 504})
-        void shouldRetryOnRetryableStatus(int status) throws Exception {
-            AtomicInteger counter = new AtomicInteger();
-
-            String result = retryExecutor.supplyWithRetry(() -> {
-                if (counter.getAndIncrement() < DEFAULT_MAX_ATTEMPTS - 1) {
-                    throw feignException(status, "Retryable Error");
-                }
-                return "success";
-            }, "send", CASE_ID);
-
-            assertThat(result).isEqualTo("success");
-            assertThat(counter.get()).isEqualTo(DEFAULT_MAX_ATTEMPTS);
-        }
-
-        @ParameterizedTest(name = "should retry once on first status {0} then succeed")
-        @ValueSource(ints = {504}) // can expand if needed
-        void shouldRetryOnceThenSucceed(int status) throws Exception {
-            AtomicInteger counter = new AtomicInteger();
-
-            String result = retryExecutor.supplyWithRetry(() -> {
-                if (counter.getAndIncrement() == 0) {
-                    throw feignException(status, "Retryable Error");
-                }
-                return "success";
-            }, "send", CASE_ID);
-
-            assertThat(result).isEqualTo("success");
-            assertThat(counter.get()).isEqualTo(2);
-        }
-
-        @Test
-        void shouldNotRetryOnNonRetryableException() {
-            AtomicInteger counter = new AtomicInteger();
-
-            assertThatThrownBy(() ->
-                retryExecutor.supplyWithRetry(() -> {
-                    counter.incrementAndGet();
-                    throw new RuntimeException("Not retryable");
-                }, "send", CASE_ID)
-            ).isInstanceOf(RuntimeException.class);
-
-            assertThat(counter.get()).isEqualTo(1);
         }
     }
 }
